Move open in web operations to their proper location
