+++ owncloudComLibrary/src/main/java/com/owncloud/android/lib/resources/appregistry/services/AppRegistryService.kt
@@ -28,4 +28,6 @@ import com.owncloud.android.lib.resources.appregistry.responses.AppRegistryRespo
 
 interface AppRegistryService : Service {
     fun getAppRegistry(): RemoteOperationResult<AppRegistryResponse>
+
+    fun getUrlToOpenInWeb(openWebEndpoint: String, fileId: String): RemoteOperationResult<String>
 }
